Move StaticCSPHeader to util
