+++ pkg/admin/routes.go
@@ -7,7 +7,6 @@ import (
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/config/configsource"
 	"github.com/authgear/authgear-server/pkg/lib/deps"
-	"github.com/authgear/authgear-server/pkg/lib/web"
 	"github.com/authgear/authgear-server/pkg/util/httproute"
 	"github.com/authgear/authgear-server/pkg/util/httputil"
 )
@@ -22,7 +21,7 @@ func NewRouter(p *deps.RootProvider, configSource *configsource.ConfigSource, au
 
 	securityMiddleware := httproute.Chain(
 		httproute.MiddlewareFunc(httputil.StaticSecurityHeaders),
-		web.StaticCSPMiddleware{
+		httputil.StaticCSPHeader{
 			CSPDirectives: []string{
 				"script-src 'self' 'unsafe-inline' cdn.jsdelivr.net",
 				"object-src 'none'",
